pyproject.toml
[project]
name = "mcp-server-docker"
version = "0.1.0"
description = "A Docker MCP Server"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docker>=7.1.0",
    "mcp>=1.1.0",
    "pydantic-settings>=2.6.1",
    "python-dotenv>=1.0.1",
]

[[project.authors]]
name = "Christian Kreiling"
email = "kreiling@hey.com"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
mcp-server-docker = "mcp_server_docker:main"
